--- src/Lykke.Service.Assets.Core/Services/IShutdownManager.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Lykke.Service.Assets.Core.Services
-{
-    public interface IShutdownManager
-    {
-        Task StopAsync();
-    }
-}
