--- src/Lykke.Service.Assets.Core/Services/IStartupManager.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace Lykke.Service.Assets.Core.Services
-{
-    public interface IStartupManager
-    {
-        Task StartAsync();
-    }
-}
