+++ src/Lykke.Service.Assets.Services/ServicesModule.cs
@@ -82,16 +82,6 @@ namespace Lykke.Service.Assets.Services
                 .SingleInstance();
 
             builder
-                .RegisterType<StartupManager>()
-                .As<IStartupManager>()
-                .SingleInstance();
-
-            builder
-                .RegisterType<ShutdownManager>()
-                .As<IShutdownManager>()
-                .SingleInstance();
-
-            builder
                 .RegisterType<WatchListService>()
                 .As<IWatchListService>()
                 .SingleInstance();
