+++ src/Lykke.Service.Assets/RabbitSubscribers/Erc20ContractCreatedMessage.cs
@@ -1,4 +1,4 @@
-﻿namespace Lykke.Service.Assets.IncomingMessages
+﻿namespace Lykke.Service.Assets.RabbitSubscribers
 {
     // NOTE: This is incoming message example
     public class Erc20ContractCreatedMessage
